PHP Project “My Online book store”
• In this project, there are many functionalities like User
Registration, User Login, Add to Cart, Checkout any many other.
In this web application using database I restricted can’t create
different username with same email id if he/she is being already
registered to this Online Book Store.
Session and Cookies are implemented and accessed as per the requirement.
Guest users are also invited but if they want to buy then they have to
register after checkout is done. As there is a double verification
whether user is logged in or not.